requirements.txt
black==24.8.0
django==5.0.7
djangorestframework==3.15.2
python-dotenv==1.0.1
pre-commit==3.8.0
isort==5.10.1
